+++ tools/RAiDER/etc/entrypoint.sh
@@ -1,0 +1,4 @@
+#!/bin/bash --login
+set -e
+conda activate RAiDER
+exec raider.py "$@"